EDU-471 Move document regeneration into admin controller
