+++ src/api-clients/document-api-client.js
@@ -126,16 +126,6 @@ class DocumentApiClient {
       )
       .then(res => res.data);
   }
-
-  postDocumentRegenerationBatch() {
-    return this.httpClient
-      .post(
-        '/api/v1/docs/document-regeneration-batch',
-        null,
-        { responseType: 'json' }
-      )
-      .then(res => res.data);
-  }
 }
 
 export default DocumentApiClient;
