+++ src/server/admin-controller.js
@@ -1,6 +1,7 @@
 import PageRenderer from './page-renderer.js';
 import permissions from '../domain/permissions.js';
 import { PAGE_NAME } from '../domain/page-name.js';
+import BatchService from '../services/batch-service.js';
 import SettingService from '../services/setting-service.js';
 import StorageService from '../services/storage-service.js';
 import DocumentService from '../services/document-service.js';
@@ -8,12 +9,13 @@ import needsPermission from '../domain/needs-permission-middleware.js';
 import ClientDataMappingService from '../services/client-data-mapping-service.js';
 
 class AdminController {
-  static get inject() { return [SettingService, DocumentService, StorageService, ClientDataMappingService, PageRenderer]; }
+  static get inject() { return [SettingService, DocumentService, StorageService, BatchService, ClientDataMappingService, PageRenderer]; }
 
-  constructor(settingService, documentService, storageService, clientDataMappingService, pageRenderer) {
+  constructor(settingService, documentService, storageService, batchService, clientDataMappingService, pageRenderer) {
     this.settingService = settingService;
     this.documentService = documentService;
     this.storageService = storageService;
+    this.batchService = batchService;
     this.clientDataMappingService = clientDataMappingService;
     this.pageRenderer = pageRenderer;
   }
@@ -29,11 +31,26 @@ class AdminController {
     return this.pageRenderer.sendPage(req, res, PAGE_NAME.admin, initialState);
   }
 
-  registerPages(app) {
-    app.get(
+  async handlePostDocumentRegenerationRequest(req, res) {
+    const { user } = req;
+    const batch = await this.batchService.createDocumentRegenerationBatch(user);
+
+    return res.status(201).send(batch);
+  }
+
+  registerPages(router) {
+    router.get(
       '/admin',
       needsPermission(permissions.ADMIN),
       (req, res) => this.handleGetAdminPage(req, res)
+    );
+  }
+
+  registerApi(router) {
+    router.post(
+      '/api/v1/admin/document-regeneration',
+      needsPermission(permissions.REGENERATE_DOCS),
+      (req, res) => this.handlePostDocumentRegenerationRequest(req, res)
     );
   }
 }
