+++ src/services/batch-service.js
@@ -4,6 +4,7 @@ import uniqueId from '../utils/unique-id.js';
 import TaskStore from '../stores/task-store.js';
 import LockStore from '../stores/lock-store.js';
 import BatchStore from '../stores/batch-store.js';
+import DocumentStore from '../stores/document-store.js';
 import { BATCH_TYPE, TASK_TYPE } from '../domain/constants.js';
 import TransactionRunner from '../stores/transaction-runner.js';
 
@@ -15,14 +16,15 @@ const CONCURRENT_IMPORT_BATCH_ERROR_MESSAGE = 'Cannot create a new import batch 
 
 class BatchService {
   static get inject() {
-    return [TransactionRunner, BatchStore, TaskStore, LockStore];
+    return [TransactionRunner, BatchStore, TaskStore, LockStore, DocumentStore];
   }
 
-  constructor(transactionRunner, batchStore, taskStore, lockStore) {
+  constructor(transactionRunner, batchStore, taskStore, lockStore, documentStore) {
     this.transactionRunner = transactionRunner;
     this.batchStore = batchStore;
     this.taskStore = taskStore;
     this.lockStore = lockStore;
+    this.documentStore = documentStore;
   }
 
   async createImportBatch({ importSource, documentsToImport, user }) {
@@ -83,6 +85,43 @@ class BatchService {
     } finally {
       await this.lockStore.releaseLock(lock);
     }
+
+    return batch;
+  }
+
+  async createDocumentRegenerationBatch(user) {
+    const existingActiveBatch = await this.batchStore.getUncompleteBatchByType(BATCH_TYPE.documentRegeneration);
+
+    if (existingActiveBatch) {
+      throw new BadRequest('Another document regeneration batch is already in progress');
+    }
+
+    const batch = {
+      _id: uniqueId.create(),
+      createdBy: user._id,
+      createdOn: new Date(),
+      completedOn: null,
+      batchType: BATCH_TYPE.documentRegeneration,
+      batchParams: {},
+      errors: []
+    };
+
+    const allDocumentKeys = await this.documentStore.getAllDocumentKeys();
+    const tasks = allDocumentKeys.map(key => ({
+      _id: uniqueId.create(),
+      batchId: batch._id,
+      taskType: TASK_TYPE.documentRegeneration,
+      processed: false,
+      attempts: [],
+      taskParams: {
+        key
+      }
+    }));
+
+    await this.transactionRunner.run(async session => {
+      await this.batchStore.createBatch(batch, { session });
+      await this.taskStore.addTasks(tasks, { session });
+    });
 
     return batch;
   }
