+++ src/services/document-service.spec.js
@@ -1,18 +1,14 @@
 /* eslint-disable max-lines */
 import sinon from 'sinon';
-import httpErrors from 'http-errors';
 import uniqueId from '../utils/unique-id.js';
 import Database from '../stores/database.js';
 import cloneDeep from '../utils/clone-deep.js';
-import TaskStore from '../stores/task-store.js';
 import LockStore from '../stores/lock-store.js';
 import DocumentService from './document-service.js';
-import { BATCH_TYPE, DOCUMENT_ORIGIN, TASK_TYPE } from '../domain/constants.js';
+import { DOCUMENT_ORIGIN } from '../domain/constants.js';
 import { SOURCE_TYPE as IMAGE_SOURCE_TYPE } from '../plugins/image/constants.js';
 import { SOURCE_TYPE as VIDEO_SOURCE_TYPE } from '../plugins/video/constants.js';
 import { createTestDocument, createTestRevisions, destroyTestEnvironment, pruneTestEnvironment, setupTestEnvironment, setupTestUser } from '../test-helper.js';
-
-const { BadRequest } = httpErrors;
 
 const createDefaultSection = () => ({
   key: uniqueId.create(),
@@ -28,7 +24,6 @@ describe('document-service', () => {
   let container;
   let user;
 
-  let taskStore;
   let lockStore;
 
   let sut;
@@ -38,7 +33,6 @@ describe('document-service', () => {
     container = await setupTestEnvironment();
     user = await setupTestUser(container);
 
-    taskStore = container.get(TaskStore);
     lockStore = container.get(LockStore);
 
     sut = container.get(DocumentService);
@@ -1219,65 +1213,6 @@ describe('document-service', () => {
         const results = await sut.getDocumentsMetadataByTags('-cretu');
 
         expect(results).toHaveLength(0);
-      });
-    });
-  });
-
-  describe('createDocumentRegenerationBatch', () => {
-    let testDocument;
-    beforeEach(async () => {
-      testDocument = await createTestDocument(container, user, {
-        title: 'Title',
-        description: 'Description',
-        slug: 'my-doc',
-        language: 'en',
-        sections: [],
-        tags: ['tag-1']
-      });
-    });
-
-    describe('when a running batch already exists', () => {
-      beforeEach(async () => {
-        await sut.createDocumentRegenerationBatch(user);
-      });
-
-      it('should return null', async () => {
-        await expect(() => sut.createDocumentRegenerationBatch(user)).rejects.toThrow(BadRequest);
-      });
-    });
-
-    describe('when a new batch is created', () => {
-      let batch = null;
-      beforeEach(async () => {
-        batch = await sut.createDocumentRegenerationBatch(user);
-      });
-
-      it('should create the new batch', () => {
-        expect(batch).toEqual({
-          _id: expect.stringMatching(/\w+/),
-          createdBy: user._id,
-          createdOn: now,
-          completedOn: null,
-          batchType: BATCH_TYPE.documentRegeneration,
-          batchParams: {},
-          errors: []
-        });
-      });
-
-      it('should create the according tasks', async () => {
-        const tasks = await taskStore.getTasksByBatchId(batch._id);
-        expect(tasks).toEqual([
-          {
-            _id: expect.stringMatching(/\w+/),
-            batchId: batch._id,
-            taskType: TASK_TYPE.documentRegeneration,
-            processed: false,
-            attempts: [],
-            taskParams: {
-              key: testDocument.key
-            }
-          }
-        ]);
       });
     });
   });
